+++ src/variables/MobileDetectVariable.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace superbig\mobiledetect\variables;
 
+use BadMethodCallException;
 use superbig\mobiledetect\MobileDetect;
 
 /**
@@ -14,6 +15,11 @@ use superbig\mobiledetect\MobileDetect;
  *   {{ craft.mobileDetect.isTablet }}
  *   {{ craft.mobileDetect.isPhone }}
  *   {{ craft.mobileDetect.is('iOS') }}
+ *   {{ craft.mobileDetect.isiOS }}
+ *   {{ craft.mobileDetect.isAndroidOS }}
+ *   {{ craft.mobileDetect.isChrome }}
+ *
+ * Any `is*` call is forwarded to the library's rule matching via __call().
  */
 class MobileDetectVariable
 {
@@ -55,5 +61,22 @@ class MobileDetectVariable
     public function getUserAgent(): ?string
     {
         return MobileDetect::getInstance()->mobileDetectService->getUserAgent();
+    }
+
+    /**
+     * Forward any `is*` call to the library's rule matching.
+     *
+     * Supports all OS, browser, and device rules:
+     *   isiOS(), isAndroidOS(), isChrome(), isSafari(), etc.
+     *
+     * @param array<mixed> $arguments
+     */
+    public function __call(string $name, array $arguments): bool
+    {
+        if (str_starts_with($name, 'is')) {
+            return $this->is(substr($name, 2));
+        }
+
+        throw new BadMethodCallException("No such method: $name");
     }
 }